refactor: co-locate Lzip/Xz stream classes with their format modules
Move LzipDecompressorStream into lzip_stream.py and XzDecompressorStream into
xz_stream.py, next to the parsing primitives and state machines they use. This
makes each format module self-contained (parsing + state machine + seekable
stream) and inverts the dependency cleanly: decompressor_stream.py is now pure
generic infrastructure (SeekPoint, DecompressorStream, _SegmentedDecompressorStream,
and the trivial zlib/brotli wrappers) that the format modules depend on, rather
than importing the format internals itself.

Also simplify the XZ seek-point round-trip: store the _XzBlockBounds object
directly in SeekPoint.state instead of packing/unpacking a
(check, unpadded_size, uncompressed_size) tuple. _make_decompressor now reads
point.state as a bounds object directly, dropping ~14 lines of field repacking.

The state machines (_LzipState / _XzState / _XzBlockChain) stay as separate
classes — forward decoding and seekable-index management are distinct
responsibilities — but now live in the same file as their stream class.

No behavior change; full suite green.
