+++ src/archivey/formats/decompressor_stream.py
@@ -21,13 +21,6 @@ from typing import (
 from typing_extensions import Buffer
 
 from archivey.exceptions import ArchiveCorruptedError, ArchiveEOFError
-from archivey.formats.lzip_stream import _LzipState, _read_index_backwards
-from archivey.formats.xz_stream import (
-    _read_xz_index_backwards,
-    _XzBlockBounds,
-    _XzBlockChain,
-    _XzState,
-)
 from archivey.internal.io_helpers import ensure_bufferedio
 
 if TYPE_CHECKING:
@@ -431,157 +424,6 @@ class _SegmentedDecompressorStream(DecompressorStream[_SDT]):
         ]
         total: "int | None" = bounds[-1].decompressed_end if bounds else None
         return points, total
-
-
-class LzipDecompressorStream(_SegmentedDecompressorStream[_LzipState]):
-    """Seekable lzip decompressor backed by Python's stdlib lzma.
-
-    Builds a seek-point table from member headers/trailers:
-      - Progressively as members are decoded during forward reads.
-      - On-demand via a one-shot backwards trailer scan (_build_index).
-
-    The table enables efficient SEEK_END (no decompression), backward seeks
-    (jump to the nearest indexed member), and forward seeks across already-
-    indexed members.
-    """
-
-    def _make_decompressor(self, point: SeekPoint) -> _LzipState:
-        return _LzipState()
-
-    def _on_completed_segments(self, units: list[tuple[int, int]]) -> None:
-        for decompressed_size, compressed_size in units:
-            self.add_seek_points([SeekPoint(self._decomp_cursor, self._comp_cursor)])
-            self._comp_cursor += compressed_size
-            self._decomp_cursor += decompressed_size
-
-    def _build_index(self, last_known: SeekPoint) -> tuple[list[SeekPoint], int | None]:
-        """Scan member trailers backwards to build the complete index.
-
-        On failure (e.g. trailing data after the last member, which is valid
-        per the lzip spec §7), logs a warning and falls back to sequential
-        decompression.
-        """
-        return self._build_index_backwards(
-            last_known,
-            _read_index_backwards,
-            lambda m: SeekPoint(m.decompressed_start, m.compressed_start),
-            "Lzip backwards index scan failed (the file may have trailing "
-            "data after the last member, which is valid per the lzip spec); "
-            "falling back to sequential decompression. Reason: %s",
-        )
-
-
-class XzDecompressorStream(_SegmentedDecompressorStream["_XzState | _XzBlockChain"]):
-    """Seekable XZ decompressor backed by Python's stdlib lzma.
-
-    Builds a block-level seek-point table:
-      - Progressively as XZ streams complete during forward reads.
-      - On-demand via _build_index (triggered by SEEK_END or forward seek past frontier).
-
-    SeekPoint.state == None  →  stream-level decompressor (_XzState), used only for the
-                                initial SeekPoint(0, 0) before any index is known.
-    SeekPoint.state == tuple →  block metadata (check, unpadded_size, uncompressed_size);
-                                uses _XzBlockChain for block-level random access.
-    """
-
-    def _make_decompressor(self, point: SeekPoint) -> "_XzState | _XzBlockChain":
-        if point.state is None:
-            return _XzState()
-        check, unpadded_size, uncompressed_size = point.state
-        start_block = _XzBlockBounds(
-            compressed_start=point.compressed_offset,
-            decompressed_start=point.decompressed_offset,
-            unpadded_size=unpadded_size,
-            uncompressed_size=uncompressed_size,
-            check=check,
-        )
-        subsequent = [
-            sp for sp in self._seek_points
-            if sp.decompressed_offset > point.decompressed_offset and sp.state is not None
-        ]
-        extra_blocks = [
-            _XzBlockBounds(
-                compressed_start=sp.compressed_offset,
-                decompressed_start=sp.decompressed_offset,
-                unpadded_size=sp.state[1],
-                uncompressed_size=sp.state[2],
-                check=sp.state[0],
-            )
-            for sp in subsequent
-        ]
-        return _XzBlockChain([start_block] + extra_blocks, cast("BinaryIO", self._inner))
-
-    def _on_completed_segments(self, units: list[tuple[int, int]]) -> None:
-        if isinstance(self._decompressor, _XzState):
-            self._update_index(units)
-        else:
-            for decomp_size, comp_size in units:
-                self._comp_cursor += comp_size
-                self._decomp_cursor += decomp_size
-
-    def _update_index(self, new_streams: list[tuple[int, int]]) -> None:
-        """Extend seek points with newly completed streams (only called for _XzState).
-
-        For each completed stream: adds a stream-level SeekPoint, then scans that
-        stream's compressed range backwards to populate block-level seek points.
-        """
-        for decompressed_size, compressed_size in new_streams:
-            stream_comp_start = self._comp_cursor
-            stream_decomp_start = self._decomp_cursor
-
-            # Skip for stream 0 — SeekPoint(0, 0) already covers it
-            if stream_decomp_start > 0:
-                self.add_seek_points(
-                    [SeekPoint(stream_decomp_start, stream_comp_start, state=None)]
-                )
-
-            stream_comp_end = stream_comp_start + compressed_size
-
-            if not self._index_built and self._inner.seekable():
-                saved_pos = self._inner.tell()
-                try:
-                    blocks = _read_xz_index_backwards(
-                        cast("BinaryIO", self._inner),
-                        stream_comp_end,
-                        stop_at=stream_comp_start,
-                        start_decompressed_offset=stream_decomp_start,
-                    )
-                    block_points = [
-                        SeekPoint(
-                            b.decompressed_start,
-                            b.compressed_start,
-                            state=(b.check, b.unpadded_size, b.uncompressed_size),
-                        )
-                        for b in blocks
-                        if b.decompressed_start > stream_decomp_start
-                    ]
-                    if block_points:
-                        self.add_seek_points(block_points)
-                except ArchiveCorruptedError as e:
-                    logger.warning(
-                        "XZ per-stream backward scan failed, block-level seek points "
-                        "for this stream will not be available: %s",
-                        e,
-                    )
-                finally:
-                    self._inner.seek(saved_pos)
-
-            self._comp_cursor = stream_comp_end
-            self._decomp_cursor += decompressed_size
-
-    def _build_index(self, last_known: SeekPoint) -> tuple[list[SeekPoint], int | None]:
-        """Full backwards scan from EOF to last_known, building block seek points."""
-        return self._build_index_backwards(
-            last_known,
-            _read_xz_index_backwards,
-            lambda b: SeekPoint(
-                b.decompressed_start,
-                b.compressed_start,
-                state=(b.check, b.unpadded_size, b.uncompressed_size),
-            ),
-            "XZ backwards index scan failed; falling back to sequential "
-            "decompression. Reason: %s",
-        )
 
 
 class ZlibDecompressorStream(DecompressorStream):
